Draft: makeLayer, small adjustments to keep compatibility
This follows from the migration of the Layer code
done in 08c86ced10.

In the older `makeLayer` function, the parameters `drawstyle`
and `transparency` could be `None`; in the new `make_layer` function,
`drawstyle` must be an explicit string, and `transparency`
must be a number. If not provided, they use default values.