Surface: upgraded Surface icons
The `Surface_Workbench` icon is the same as `Surface_Surface`,
which is now an SVG and not XPM.
The lines are thicker and the colors have been adjusted
with a gradient, instead of being solid and plain looking.

Also the following were updated: `BSplineSurface`,
`BezierSurface`, `Cut`, `Filling`, `Sewing.

The `BezierSurface`, `Cut`, and `Sewing` icons are not currently used
in the workbench, but they were still updated in case
a future command uses them.